extract MovieInfo to its own class
